+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Watchman
 
+* *[v0.2.0-rc1]*
+  * [FEATURE]: Added functionality to campaign_email action to send receipt emails for `paymentMade` event
+
 ## v0.1.0 (April 4, 2016)
 * *[v0.1.0-rc1]*
   * [FIX]: Fix for an issue where some campaigns may fail to be fetched